--- a/langcorrect/contributions/models.py
+++ b/langcorrect/contributions/models.py
@@ -9,7 +9,7 @@ class Contribution(SoftDeletableModel, TimeStampedModel):
     class Meta:
         ordering = ["-total_points"]
 
-    user = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
+    user = models.OneToOneField(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
     total_points = models.IntegerField(default=0)
     post_count = models.IntegerField(default=0)
     correction_count = models.IntegerField(default=0)