--- a/langcorrect/templates/corrections/popular_correctors.html
+++ b/langcorrect/templates/corrections/popular_correctors.html
@@ -2,7 +2,10 @@
 {% load static %}
 
 <div class="card">
-  <div class="card-header bg-transparent text-center">{% translate "Popular Correctors" %}</div>
+  <div class="card-header bg-transparent text-center">
+    <div>{% translate "Popular Correctors" %}</div>
+    <small class="text-muted">*updates every 5 minutes</small>
+  </div>
   <nav>
     <div class="nav nav-tabs justify-content-center"
          id="nav-tab"