--- a/langcorrect/templates/navbar.html
+++ b/langcorrect/templates/navbar.html
@@ -1,6 +1,8 @@
 {% load i18n %}
 {% load static %}
+{% load notifications_tags %}
 
+{% notifications_unread as unread_count %}
 <nav class="navbar navbar-expand-md navbar-light bg-white border-bottom">
   <div class="container  align-items-center">
     <a class="navbar-brand" href="{% url 'home' %}">
@@ -16,11 +18,11 @@
       </a>
       {% if request.user.is_authenticated %}
         <button id="notification-button"
-                class="btn {% if request.user.notifications.unread.count %} btn-primary {% else %} text-muted {% endif %} me-2 d-md-none"
+                class="btn {% if unread_count %} btn-primary {% else %} text-muted {% endif %} me-2 d-md-none"
                 data-bs-toggle="modal"
                 data-bs-target="#mobileNotifications">
           <i class="fas fa-bell fa-fw"></i>
-          {{ request.user.notifications.unread.count }}
+          {{ unread_count }}
         </button>
       {% endif %}
       <button class="navbar-toggler p-2"
@@ -54,9 +56,6 @@
             <li>
               <a class="dropdown-item px-3 py-2" href="{% url 'rankings' %}"><i class="fas fa-trophy fa-fw"></i> {% translate "Rankings" %}</a>
             </li>
-            {#            <li>#}
-            {#              <a class="dropdown-item" href=""><i class="fas fa-hiking fa-fw"></i> {% translate "Challenges" %}</a>#}
-            {#            </li>#}
           </ul>
         </li>
       </ul>
@@ -77,11 +76,11 @@
           </li>
           <li class="d-none d-md-block">
             <button id="notification-button"
-                    class="btn {% if request.user.notifications.unread.count %} btn-primary {% else %} text-muted {% endif %} me-2"
+                    class="btn {% if unread_count %} btn-primary {% else %} text-muted {% endif %} me-2"
                     data-bs-toggle="modal"
                     data-bs-target="#mobileNotifications">
               <i class="fas fa-bell fa-fw"></i>
-              {{ request.user.notifications.unread.count }}
+              {{ unread_count }}
             </button>
           </li>
           <li class="dropdown d-inline-block mt-2 mt-md-0">