--- a/langcorrect/templates/posts/partials/post_card.html
+++ b/langcorrect/templates/posts/partials/post_card.html
@@ -26,9 +26,9 @@
         <a href="{{ post.get_absolute_url }}"
            class="stretched-link link-secondary"></a>
       {% endif %}
-      {% if post.tags.names %}
+      {% if post.tags.all %}
         <div class="mt-3">
-          {% for tag in post.tags.names %}<span class="badge bg-secondary me-1">{{ tag }}</span>{% endfor %}
+          {% for tag in post.tags.all %}<span class="badge bg-secondary me-1">{{ tag.name }}</span>{% endfor %}
         </div>
       {% endif %}
     </div>