This PR significantly optimizes the performance and page load speed by addressing n+1 query issues and implementing caching for popular correctors.

**Old**
page load speed: ~800ms
default 303.54 ms (273 queries including 267 similar and 183 duplicates ) 

**New**
page load speed: ~255ms
default 65.01 ms (141 queries including 132 similar and 80 duplicates )